Split reference document storage tests
Move shared reference-document storage builders into a reusable fixture
module and separate binding repository scenarios from the document
repository tests.

Remove the file-level `too-many-lines` suppression from the original test
module while keeping each focused module below the configured line limit.
